feat(Admin): Split up the admin profiles components
